init: 환경 설정 완료 ( Parcel + Typescript )
