--- src/css/shared/index.css
@@ -1,9 +1,0 @@
-@import "modules/button.css";
-@import "modules/layout.css";
-@import "modules/sizing.css";
-@import "modules/typhography.css";
-@import "modules/spacing.css";
-@import "modules/interactivity.css";
-@import "modules/border.css";
-@import "modules/flex.css";
-@import "modules/background.css";
