--- src/css/ui/index.css
@@ -1,3 +1,0 @@
-@import 'modules/modal.css';
-@import 'modules/chip.css';
-@import 'modules/youtubeCard.css';
