+++ step1/README.md
@@ -1,5 +1,5 @@
 <p align="middle" >
-  <img width="200px;" src="./src/images/readme/laptop_with_youtube_logo.png"/>
+  <img width="200px;" src="srcmages/readme/laptop_with_youtube_logo.png"/>
 </p>
 <h2 align="middle">level1 - 나만의 유튜브 강의실</h2>
 <p align="middle">자바스크립트와 외부 API를 이용해 구현 하는 나만의 유튜브 강의실</p>
@@ -18,7 +18,7 @@
 ## 🔥 Projects!
 
 <p align="middle">
-  <img src="./src/images/readme/youtube_classroom_preview.png">
+  <img src="srcmages/readme/youtube_classroom_preview.png">
 </p>
 
 ### 🎯 step1 검색 기능
