+++ step1/src/index.html
@@ -4,7 +4,6 @@
     <meta charset="UTF-8" />
     <title>👩🏻‍💻 유튜브 강의실</title>
     <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <link rel="stylesheet" href="./src/css/index.css" />
   </head>
   <body>
     <div id="app">
@@ -119,6 +118,6 @@
         </div>
       </div>
     </div>
-    <script type="module" src="./src/js/index.js"></script>
+    <script src="./index.ts"></script>
   </body>
 </html>
